+++ Assets/Scripts/data/include/LearnsetMove.cs
@@ -1,8 +1,8 @@
 public readonly struct LearnsetMove
 {
-    public readonly byte level;
+    public readonly int level;
     public readonly MoveID move;
-    public LearnsetMove(byte Level, MoveID Move)
+    public LearnsetMove(int Level, MoveID Move)
     {
         level = Level;
         move = Move;
