+++ Assets/Scripts/src/include/Side.cs
@@ -1,13 +1,13 @@
 public class Side
 {
     public bool lightScreen = false;
-    public byte lightScreenTurns = 0;
+    public int lightScreenTurns = 0;
     public bool reflect = false;
-    public byte reflectTurns = 0;
+    public int reflectTurns = 0;
     public bool auroraVeil = false;
-    public byte auroraVeilTurns = 0;
+    public int auroraVeilTurns = 0;
     public bool mist = false;
-    public byte mistTurns = 0;
+    public int mistTurns = 0;
     public int spikes = 0;
     public bool safeguard = false;
     public int safeguardTurns = 0;
